+++ src/functions/mod.rs
@@ -1,3 +1,4 @@
 pub mod list;
 pub mod show;
 pub mod position;
+pub mod set_mode;