+++ src/main.rs
@@ -17,8 +17,8 @@ mod wlr_client;
 use clap::Parser;
 use cli::Cli;
 use cli::Commands;
+use cli::Property;
 use functions::position::TargetSetup;
-use wlr_client::wlr_mode::OutputMode;
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     #[cfg(debug_assertions)]
@@ -57,7 +57,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
                 .get(&reference)
                 .cloned()
                 .ok_or("reference output does not exist")?;
-            Ok(functions::position::exec(
+            functions::position::exec(
                 &TargetSetup {
                     target: target_head,
                     reference: reference_head,
@@ -65,7 +65,8 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
                     alignment,
                 },
                 &client,
-            )?)
+            )?;
+            Ok(())
         }
         Commands::Set {
             target,
@@ -76,16 +77,14 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
                 .get(&target)
                 .cloned()
                 .ok_or("target output does not exist")?;
-            let mut modes: Vec<&OutputMode> = target_head.mode_ids().iter()
-                .map(|id| configs.get_mode(id).expect("Unexpected error"))
-                .collect();
-            modes.sort_by(|a, b| b.cmp(a));
-
-            let desired: usize = value.trim().parse().expect("Expected integer identifier for the mode");
-            if let Some(target_mode) = modes.get(desired) {
-                println!("You selected {} @ {} mode", target_mode.size(), target_mode.refresh());
-            } else {
-                return Err("Invalid mode identifier")?;
+            match property {
+                Property::Mode => {
+                    let desired: usize = value
+                        .trim()
+                        .parse()
+                        .expect("Expected integer identifier for the mode");
+                    functions::set_mode::exec(&target_head, desired, &client)?;
+                }
             }
 
             Ok(())